pyproject.toml
[tool.poetry]
name = "API-exp"
version = "0.1.0"
description = "Experiment with the api"
authors = ["alekspletnev <alekspletnev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
django = "^4.1.3"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
